@@ -16,16 +16,18 @@
     "lodash": "^4.16.4",
     "lowdb": "^0.14.0",
     "materialize-css": "^0.97.7",
-    "react": "^15.3.2",
+    "react": "^15.4.1",
+    "react-addons-css-transition-group": "^15.4.1",
     "react-dom": "^15.3.2",
     "react-icons": "^2.2.1",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.8.1",
-    "react-router-redux": "^4.0.6",
+    "react-redux": "^4.4.6",
+    "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.7",
     "react-tooltip": "^3.2.1",
     "redux": "^3.6.0",
     "redux-form": "^6.1.1",
     "redux-logger": "^2.7.0",
+    "redux-promise-middleware": "^4.1.0",
     "redux-thunk": "^2.1.0",
     "uuid": "^2.0.3",
     "webpack": "^1.13.2"
