@@ -1,20 +1,17 @@
-import * as types from '../constants/ActionTypes';
+import * as types from '../constants/GameSystemActionTypes';
+import GameSystemController from '../models/game_system_data';
 
-function LoadGameSystems(game_systems) {
+function loadGameSystems(game_systems) {
   return {
     type: types.LOAD_GAME_SYSTEMS,
     game_systems: game_systems
   }
 }
 
-export function CollectGameSystems(){
-  var game_systems = [
-    {
-      id: 1,
-      gameSystemName: "Dungeons & Dragons"
-    }
-  ];
+export function LoadGameSystems(){
   return dispatch => {
-    dispatch(LoadGameSystems(game_systems))
+    GameSystemController.getGameSystems(game_systems => {
+      dispatch(loadGameSystems(game_systems));
+    })
   }
 }
