@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 
-
 const CharacterListItem = (props) => {
   let { selected } = props.state;
   let selectedCharacterClass = 'selectedCharacter';
@@ -9,11 +8,11 @@ const CharacterListItem = (props) => {
       onClick={() =>{props._selectCharacter({selected: c.id})}}
       key={c.id}
       className={`character ${selected == c.id ? selectedCharacterClass : null}`} >
-      <div className="characterAvatar" style={{backgroundImage: `url('./${c.characterAvatarUri}')`}}></div>
+      <div className="characterAvatar" style={{backgroundImage: `url('${c.AvatarUri}')`}}></div>
       <div className="characterDetail">
-        <p className="characterName">{c.characerName}</p>
-        <p>{c.characerProfession}</p>
-        <p  className="opt-sm">{ campaignName(c.id) }</p>
+        <p className="characterName">{c.characterName}</p>
+        <p>{c.characterProfession}</p>
+        <p className="opt-sm">{ campaignName(c.id) }</p>
       </div>
       <div className="expGauge">
         <span className="completion"></span>
