@@ -1,22 +1,12 @@
 import React from 'react';
 
 import AppCtrl from './appctrl';
-import Navigation from '../containers/Navigation';
+import {Navigation} from './navigation';
 import PlayerCtrl from './playerctrl';
 
 const DefaultHeader = (props) => {
-  const _renderNavigation = () => {
-    return (
-      <Navigation {...props} />
-    )
-  }
-
-  const _renderPlayerCtrl = () => {
-    return (
-      <PlayerCtrl {...props} />
-    )
-  }
-
+  const _renderNavigation = () => <Navigation {...props} />
+  const _renderPlayerCtrl = () => <PlayerCtrl {...props} />
   return (
     <div className={`ApplicationHeader header-scroll small`} >
       <AppCtrl {...props} />
