@@ -2,150 +2,28 @@ import React, { Component } from 'react';
 import { ipcRenderer, remote } from 'electron';
 import { connect}  from 'react-redux';
 import { IoChevronLeft, IoChevronRight,IoNavicon } from 'react-icons/lib/io';
-
-
-const ChatDisplay = (props) => {
-  return (
-    <div className='Chat Module '>
-    <hgroup className='ModuleHeader deep-purple'>
-      <IoNavicon size={28} className={`HamburgNav`} color={`rgba(255,255,255,1)`}/>
-      <h3>Conversations</h3>
-    </hgroup>
-      <div className="chatContainer">
-        <div className={`ChatBody scroll3`}>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
+import { ChatBubble } from '../components/chatbubble.js'
+class ChatDisplay extends Component {
+  constructor (props) {
+    super(props);
+  }
+  render(){
+    let {conversations} = this.props;
+    return (
+      <div className='Chat Module '>
+        <hgroup className='ModuleHeader deep-purple'>
+          <h3>Conversations</h3>
+          <IoNavicon size={28} className={`pull-right HamburgNav`} color={`rgba(255,255,255,1)`}/>
+        </hgroup>
+        <div className="chatContainer">
+          <div className={`ChatBody scroll3`}>
+            {conversations.map((c,i) => <ChatBubble key={i} />)}
           </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
-          <div className={`chatBubble `}>
-            <div className={`playerAvatar`} style={{backgroundImage: `url('./images/rogue.jpg')`}}></div>
-            <div className={`playerContentHolder`} >
-              <h2>thing one test title</h2>
-              <h3>thing one test title</h3>
-            </div>
-            <IoChevronRight className={`indicator`}/>
-          </div>
-
         </div>
       </div>
-    </div>
-  );
+    );
+  }
 }
-
 
 
 export default ChatDisplay
