@@ -3,8 +3,9 @@ import { ipcRenderer, remote } from 'electron';
 import { connect}  from 'react-redux';
 import { IoChevronLeft, IoChevronRight } from 'react-icons/lib/io';
 import CharacterListModule from '../container_modules/CharacterList';
+import {LoadCharacter} from '../actions/CharacterActions';
 import ChatDisplayModule from '../container_modules/ChatDisplay';
-import CampaignDisplayModule from '../container_modules/CampaignDisplay';
+import CharacterViewer from '../container_modules/CharacterViewer';
 import GamesystemDisplayModule from '../container_modules/GamesystemDisplay';
 
 
@@ -12,33 +13,48 @@ class DefaultView extends Component {
   constructor (props) {
     super(props);
     this.state = {
-      selected: {}
+      selected_character: {}
     };
   }
   componentDidMount(){
+    let {selected_character} = this.props
+
+    this.setState({
+      selected_character: selected_character
+    })
     // alert(JSON.stringify(this.props.player));
+  }
+  componentWillReceiveProps(nextProps){
+    if(this.props.selected_character != nextProps.selected_character){
+      this.setState({
+        player: nextProps.player,
+        selected_character: nextProps.selected_character
+      })
+    }
   }
 
   _onSelectCharacter(character) {
+    let {dispatch} = this.props;
+    dispatch(LoadCharacter(character.id));
     this.setState({
-      selected: character
+      selected_character: character
     });
   }
 
   render() {
 
-    let { authenticated } = this.props;
-    let { selected } = this.state;
+    let { authenticated, player, conversations } = this.props;
+    let { selected_character } = this.state;
     const Style={"color":"#FFF", visibility: authenticated ? "visible" : "hidden" };
     return (
       <div className="DefaultView">
         <div className='DashboardModules'>
           <div style={{ flex: 1, flexDirection: 'column', display: 'flex', maxWidth: '350px'}}>
             <CharacterListModule _onSelectCharacter={( id ) => { this._onSelectCharacter(id) }}/>
-            <ChatDisplayModule />
+            <ChatDisplayModule conversations={conversations}/>
           </div>
           <div style={{ flex: 1, flexDirection: 'column', display: 'flex', maxWidth: '1140px'}}>
-            <CampaignDisplayModule selected={selected} />
+            <CharacterViewer player={player} empty={selected_character.id == undefined ? true : false} _onSelectCharacter={( id ) => { this._onSelectCharacter(id) }} selected_character={selected_character} />
             <GamesystemDisplayModule />
           </div>
         </div>
@@ -49,7 +65,10 @@ class DefaultView extends Component {
 
 const mapStateToProps = (state) => {
   return {
-    player: state.Player.player
+    conversations: state.Conversations.conversations,
+    player: state.Player.player,
+    characters: state.Characters.characters,
+    selected_character: state.Character.selected_character,
   };
 }
 export default connect(mapStateToProps)(DefaultView)
