@@ -1,21 +1,22 @@
 import React, { Component } from 'react';
-import { ipcRenderer, remote } from 'electron';
 import { connect}  from 'react-redux';
-import { IoChevronLeft, IoChevronRight } from 'react-icons/lib/io';
-import { Link } from 'react-router';
 
-class GameSystemView extends Component {
-
+class GameSystemsView extends Component {
   constructor (props) {
     super(props);
   }
 
   render() {
+    let {game_systems } = this.props;
     return (
-      <div className="GameSystemView releaseAppmargin">
-        <div className="ViewPlaceholderContainer">
-          <h3>Game System View</h3>
-          <button onClick={()=>{ alert('click Go Back') }}>Go Back</button>
+      <div className="GameSystemsView releaseAppmargin">
+        <hgroup>
+          <h3>Game Systems View</h3>
+        </hgroup>
+        <div className="Layout scroll3">
+          <div className={`ListViewMain GameSystemsMain`}>
+          {game_systems.map((g,i) => <div key={i} className={`ListViewBlock GameSystemBlock`}>{g.gameSystemName}</div>)}
+          </div>
         </div>
       </div>
     );
@@ -23,6 +24,8 @@ class GameSystemView extends Component {
 }
 
 const mapStateToProps = (state) => {
-  return {}
+  return {
+    game_systems: state.GameSystems.game_systems
+  }
 }
-export default connect(mapStateToProps)(GameSystemView)
+export default connect(mapStateToProps)(GameSystemsView)
