@@ -1,16 +1,33 @@
-import {
-  FETCH_CHARACTERS,
-  SELECT_CHARACTER
-} from '../actions/types';
+import * as types from '../constants/CharacterActionTypes';
 
 
-export function Characters(state = [], action )  {
+const CharacterState = {
+  selected_character: {},
+};
+
+const CharactersState = {
+  characters: [],
+};
+
+// CHARACTER REDUCER
+export const Character = (state = CharacterState, action ) => {
   switch (action.type) {
-    case FETCH_CHARACTERS:
-      return [
-        ...state,
-        ...action.payload,
-      ];
+    case types.SELECT_CHARACTER:
+      return Object.assign({}, state, {
+        selected_character: action.selected_character
+      });
+   default:
+    return state;
+  }
+}
+
+// CHARACTERS
+export const Characters = (state = CharactersState, action ) => {
+  switch (action.type) {
+    case types.LOAD_CHARACTERS:
+      return Object.assign({}, state, {
+        characters: action.characters
+      });
    default:
     return state;
   }
