@@ -1,20 +1,29 @@
-import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
-import { routerReducer,routerMiddleware, push} from 'react-router-redux';
+import { createStore, applyMiddleware, compose } from 'redux';
 import rootReducer from '../reducers';
 import thunk from 'redux-thunk'
 import Evolition_DnD5e from '../gamesystems/dnd5e';
+import promiseMiddleware from 'redux-promise-middleware';
 
-const initialize = {}
+export default function configureStore(initialState = {}){
+  let middleWare = [thunk, promiseMiddleware(),  Evolition_DnD5e];
+  let store;
 
-const configureStore = (initialState = initialize) => {
-
-  let middleware = applyMiddleware(thunk, Evolition_DnD5e);
-
-  if(process.env.NODE_ENV === 'development') {
-    middleware = compose(middleware);
+  if (process.env.NODE_ENV === 'development') {
+    store = createStore(rootReducer, initialState, compose(
+      applyMiddleware(...middleWare),
+      typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
+    ));
+  } else {
+    store = createStore(rootReducer, initialState, compose(applyMiddleware(...middleWare), f => f));
   }
-  const store = middleware(createStore)(rootReducer, initialState);  // reducers, initialState);
+
+  if (module.hot) {
+    // Enable Webpack hot module replacement for reducers
+    module.hot.accept('reducers', () => {
+      const nextReducer = require('../reducers');
+      store.replaceReducer(nextReducer);
+    });
+  }
+
   return store;
 };
-
-export default configureStore
